requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
eventlet==0.33.0
SQLAlchemy==1.4.23
opencv-python==4.5.3.56
pytesseract==0.3.8
ultralytics==8.0.0
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0 